factor out caching logic from vtable code.
introduces separation of "logic" which makes testing simpler and working with LLMs which are choking on larger files because the context gets polluted.
