+++ packages/lix-sdk/src/state/cache/clear-state-cache.ts
@@ -1,11 +1,11 @@
 import type { Kysely } from "kysely";
-import type { Lix } from "../lix/open-lix.js";
-import type { LixInternalDatabaseSchema } from "../database/schema.js";
+import type { Lix } from "../../lix/open-lix.js";
+import type { LixInternalDatabaseSchema } from "../../database/schema.js";
 
 /**
  * Clears the internal state cache.
  */
-export async function clearCache(args: { lix: Lix }): Promise<void> {
+export async function clearStateCache(args: { lix: Lix }): Promise<void> {
 	const executeInTransaction = async (trx: Lix["db"]) => {
 		await (trx as unknown as Kysely<LixInternalDatabaseSchema>)
 			.deleteFrom("internal_state_cache")
