+++ packages/lix-sdk/src/state/schema.ts
@@ -14,6 +14,8 @@ import {
 } from "./materialize-state.js";
 import { commitDeterminsticSequenceNumber } from "../deterministic/sequence.js";
 import { timestamp } from "../deterministic/timestamp.js";
+import { applyStateCacheSchema } from "./cache/schema.js";
+import { selectFromStateCache } from "./cache/select-from-state-cache.js";
 
 // Virtual table schema definition
 const VTAB_CREATE_SQL = `CREATE TABLE x(
@@ -38,6 +40,7 @@ export function applyStateDatabaseSchema(
 	hooks: LixHooks
 ): SqliteWasmDatabase {
 	applyMaterializeStateSchema(sqlite);
+	applyStateCacheSchema({ sqlite });
 
 	sqlite.createFunction({
 		name: "validate_snapshot_content",
@@ -429,7 +432,7 @@ export function applyStateDatabaseSchema(
 				}
 
 				// Try cache first - include inherited entities via union
-				const cacheResults = selectFromCache(sqlite, filters);
+				const cacheResults = selectFromStateCache(sqlite, filters);
 
 				cursorState.results = cacheResults || [];
 				cursorState.rowIndex = 0;
@@ -549,7 +552,7 @@ export function applyStateDatabaseSchema(
 					// Re-query cache after population
 
 					// Re-query after population with inheritance logic
-					const newResults = selectFromCache(sqlite, filters);
+					const newResults = selectFromStateCache(sqlite, filters);
 					cursorState.results = newResults || [];
 				} else {
 					if (canLog()) {
@@ -947,23 +950,6 @@ export function applyStateDatabaseSchema(
 
 	// Create the cache table for performance optimization and the untracked state table
 	const sql = `
-  CREATE TABLE IF NOT EXISTS internal_state_cache (
-    entity_id TEXT NOT NULL,
-    schema_key TEXT NOT NULL,
-    file_id TEXT NOT NULL,
-    version_id TEXT NOT NULL,
-    plugin_key TEXT NOT NULL,
-    snapshot_content TEXT, -- Allow NULL for deletions
-    schema_version TEXT NOT NULL,
-    created_at TEXT NOT NULL,
-    updated_at TEXT NOT NULL,
-    inherited_from_version_id TEXT,
-    inheritance_delete_marker INTEGER DEFAULT 0, -- Flag for copy-on-write deletion markers
-    change_id TEXT, -- Allow NULL during migration and for deletion markers 
-    change_set_id TEXT, -- Allow NULL until changeset is created at commit
-    PRIMARY KEY (entity_id, schema_key, file_id, version_id)
-  );
-
   -- Table for untracked state that bypasses change control
   CREATE TABLE IF NOT EXISTS internal_state_all_untracked (
     entity_id TEXT NOT NULL,
@@ -1089,171 +1075,6 @@ function getColumnName(columnIndex: number): string {
 	return columns[columnIndex] || "unknown";
 }
 
-function selectFromCache(
-	sqlite: SqliteWasmDatabase,
-	filters: Record<string, any>
-): any[] {
-	const filterBindings = Object.values(filters);
-	const buildWhereClause = (tableAlias: string = "") => {
-		const conditions: string[] = [];
-		const prefix = tableAlias ? `${tableAlias}.` : "";
-
-		Object.keys(filters).forEach((column) => {
-			conditions.push(`${prefix}${column} = ?`);
-		});
-
-		return conditions.length > 0 ? `WHERE ${conditions.join(" AND ")}` : "";
-	};
-
-	const statement = `SELECT * FROM (
-		-- 1. Untracked state (highest priority)
-		SELECT 
-			rowid,
-			entity_id, 
-			schema_key, 
-			file_id, 
-			version_id, 
-			plugin_key,
-			snapshot_content, 
-			schema_version, 
-			created_at, 
-			updated_at,
-			NULL as inherited_from_version_id, 
-			'untracked' as change_id, 
-			1 as untracked,
-			'untracked' as change_set_id
-		FROM internal_state_all_untracked
-		
-		UNION ALL
-		
-		-- 2. Tracked state (second priority) - only if no untracked exists
-		SELECT 
-			rowid,
-			entity_id, 
-			schema_key, 
-			file_id, 
-			version_id, 
-			plugin_key, 
-			snapshot_content, 
-			schema_version, 
-			created_at, 
-			updated_at,
-			inherited_from_version_id, 
-			change_id, 
-			0 as untracked,
-			change_set_id
-		FROM internal_state_cache
-		WHERE inheritance_delete_marker = 0  -- Hide copy-on-write deletions
-		AND NOT EXISTS (
-			SELECT 1 FROM internal_state_all_untracked unt
-			WHERE unt.entity_id = internal_state_cache.entity_id
-			  AND unt.schema_key = internal_state_cache.schema_key
-			  AND unt.file_id = internal_state_cache.file_id
-			  AND unt.version_id = internal_state_cache.version_id
-		)
-		
-		UNION ALL
-		
-		-- 3. Inherited tracked state (lower priority) - only if no untracked or tracked exists
-		SELECT 
-			rowid,
-			isc.entity_id, 
-			isc.schema_key, 
-			isc.file_id, 
-			vi.version_id, -- Return child version_id
-			isc.plugin_key, 
-			isc.snapshot_content, 
-			isc.schema_version, 
-			isc.created_at, 
-			isc.updated_at,
-			vi.parent_version_id as inherited_from_version_id, 
-			isc.change_id, 
-			0 as untracked,
-			isc.change_set_id
-		FROM (
-			-- Get version inheritance relationships from cache
-			SELECT 
-				json_extract(isc_v.snapshot_content, '$.id') AS version_id,
-				json_extract(isc_v.snapshot_content, '$.inherits_from_version_id') AS parent_version_id
-			FROM internal_state_cache isc_v
-			WHERE isc_v.schema_key = 'lix_version'
-		) vi
-		JOIN internal_state_cache isc ON isc.version_id = vi.parent_version_id
-		WHERE vi.parent_version_id IS NOT NULL
-		-- Only inherit entities that exist (not deleted) in parent
-		AND isc.inheritance_delete_marker = 0
-		-- Don't inherit if child has tracked state
-		AND NOT EXISTS (
-			SELECT 1 FROM internal_state_cache child_isc
-			WHERE child_isc.version_id = vi.version_id
-			  AND child_isc.entity_id = isc.entity_id
-			  AND child_isc.schema_key = isc.schema_key
-			  AND child_isc.file_id = isc.file_id
-		)
-		-- Don't inherit if child has untracked state
-		AND NOT EXISTS (
-			SELECT 1 FROM internal_state_all_untracked unt
-			WHERE unt.version_id = vi.version_id
-			  AND unt.entity_id = isc.entity_id
-			  AND unt.schema_key = isc.schema_key
-			  AND unt.file_id = isc.file_id
-		)
-		
-		UNION ALL
-		
-		-- 4. Inherited untracked state (lowest priority) - only if no untracked or tracked exists
-		SELECT 
-			rowid,
-			unt.entity_id, 
-			unt.schema_key, 
-			unt.file_id, 
-			vi.version_id, -- Return child version_id
-			unt.plugin_key, 
-			unt.snapshot_content, 
-			unt.schema_version, 
-			unt.created_at, 
-			unt.updated_at,
-			vi.parent_version_id as inherited_from_version_id, 
-			'untracked' as change_id, 
-			1 as untracked,
-			'untracked' as change_set_id
-		FROM (
-			-- Get version inheritance relationships from cache
-			SELECT 
-				json_extract(isc_v.snapshot_content, '$.id') AS version_id,
-				json_extract(isc_v.snapshot_content, '$.inherits_from_version_id') AS parent_version_id
-			FROM internal_state_cache isc_v
-			WHERE isc_v.schema_key = 'lix_version'
-		) vi
-		JOIN internal_state_all_untracked unt ON unt.version_id = vi.parent_version_id
-		WHERE vi.parent_version_id IS NOT NULL
-		-- Don't inherit if child has tracked state
-		AND NOT EXISTS (
-			SELECT 1 FROM internal_state_cache child_isc
-			WHERE child_isc.version_id = vi.version_id
-			  AND child_isc.entity_id = unt.entity_id
-			  AND child_isc.schema_key = unt.schema_key
-			  AND child_isc.file_id = unt.file_id
-		)
-		-- Don't inherit if child has untracked state
-		AND NOT EXISTS (
-			SELECT 1 FROM internal_state_all_untracked child_unt
-			WHERE child_unt.version_id = vi.version_id
-			  AND child_unt.entity_id = unt.entity_id
-			  AND child_unt.schema_key = unt.schema_key
-			  AND child_unt.file_id = unt.file_id
-		)
-	) as combined_results`;
-
-	const result = sqlite.exec({
-		sql: `${statement} ${buildWhereClause("combined_results")}`,
-		bind: [...filterBindings],
-		returnValue: "resultRows",
-	});
-
-	return result;
-}
-
 export type StateView = Omit<StateAllView, "version_id">;
 
 export type StateAllView = {
@@ -1270,23 +1091,6 @@ export type StateAllView = {
 	change_id: Generated<string>;
 	untracked: Generated<boolean>;
 	change_set_id: Generated<string>;
-};
-
-// Cache table type (internal table for state materialization)
-export type InternalStateCacheTable = {
-	entity_id: string;
-	schema_key: string;
-	file_id: string;
-	version_id: string;
-	plugin_key: string;
-	snapshot_content: string | null; // JSON string, NULL for deletions
-	schema_version: string;
-	created_at: string;
-	updated_at: string;
-	inherited_from_version_id: string | null;
-	inheritance_delete_marker: number; // 1 for copy-on-write deletion markers, 0 otherwise
-	change_id: string;
-	change_set_id: string | null;
 };
 
 export type InternalStateAllUntrackedTable = {
@@ -1306,9 +1110,6 @@ export type StateRow = Selectable<StateView>;
 export type NewStateRow = Insertable<StateView>;
 export type StateRowUpdate = Updateable<StateView>;
 
-export type StateCacheRow = Selectable<InternalStateCacheTable>;
-export type NewStateCacheRow = Insertable<InternalStateCacheTable>;
-export type StateCacheRowUpdate = Updateable<InternalStateCacheTable>;
 
 // Types for the internal_change TABLE
 export type InternalChangeInTransaction =
